Implemented the a menu and pause screen for the player to allow them to select a username and view highscores.
Highscores have been added, this means that a highscore is saved to a file on the computer with the username and score. The file for highscores is updated when the pause screen is loaded from the file.

Implemented further bug fixes to improve performance of the game.

Implemented crop fixes to improve the fluidity of the game.

Implemented a system so the Hostile can pursure a new crop once it destroys the crop it is currently targetting.